Implemented 128-bit wide MVE impl using 16x32 bit folds
The folding technique outlined in Intel's whitepaper works perfectly
fine with 16-bits, except since we are trying to use it on a 32-bit CRC,
and the folding constant is derived from the CRC's polynomial, we end
up needing to multiply the 16-bit lanes with 32-bit constants.

The 2x multiplications necessary to do with is actually a great improvement,
the alternative would be ~4x multiplications for long-multiply.

More annoying is that the results ends up being 144-bits long, which
doesn't fit well in our 128-bit registers. Fortunately MVE's extremely
handy vshlc instruction handles this perfectly, performing the required
16-bit shift and placing/consuming the 16-bit overflow in a general purpose
register.

The result is a really tight inner loop with only ~2x the overhead of
a hypothetical 32-bit polynomial multiply operation.