Added an emulated 32-bit pmul impl for comparison
It's always interesting when the results are surprising. Simply
emulating 32-bit pmuls turned out to be very close to our specialized
16-bit pmul folding impl.

The 32x32 long-multiplication requires 2x more 16-bit pmuls than the
16x32 multiplication, but makes up for this by folding 2x more data
in one emulated multiplication. So it ends up with the same number of
muls, only losing in the extra shuffling to xor the results into the
right places.